requirements.txt
influxdb == 5.3.1
python-kasa == 0.4.3

